+++ app/controllers/MessageController.scala
@@ -20,7 +20,7 @@ object MessageController extends Controller with Secured {
   
   //private final val logger: Logger = LoggerFactory.getLogger(classOf[Application])
   
-  def get(messageId: Int) = IsAuthenticated{ username => implicit request =>
+  def get(messageId: Long) = IsAuthenticated{ username => implicit request =>
     //logger.info("in MessageController.get(${messageId})")
     println("in MessageController.get(${messageId})")
     Ok("")
@@ -62,7 +62,7 @@ object MessageController extends Controller with Secured {
       }
   }
 
-  def delete(messageId: Int) = IsAuthenticated{ username => implicit request =>
+  def delete(messageId: Long) = IsAuthenticated{ username => implicit request =>
     //logger.info(s"in MessageController.delete(${messageId})")
     println(s"in MessageController.delete(${messageId})")
     
@@ -71,7 +71,7 @@ object MessageController extends Controller with Secured {
     Ok(HttpResponseUtil.success("Successfully deleted!"))
   }
   
-  def markStar(messageId: Int) = IsAuthenticated{ username => implicit request =>
+  def markStar(messageId: Long) = IsAuthenticated{ username => implicit request =>
     //logger.info(s"in MessageController.markStar(${messageId})")
     println(s"in MessageController.markStar(${messageId})")
     
@@ -79,7 +79,7 @@ object MessageController extends Controller with Secured {
     Ok(HttpResponseUtil.success("Successfully marked star!"))
   }
 
-  def removeStar(messageId: Int) = IsAuthenticated{ username => implicit request =>
+  def removeStar(messageId: Long) = IsAuthenticated{ username => implicit request =>
     //logger.info(s"in MessageController.removeStar(${messageId})")
     println(s"in MessageController.removeStar(${messageId})")
     
@@ -87,7 +87,7 @@ object MessageController extends Controller with Secured {
     Ok(HttpResponseUtil.success("Successfully remove star!"))
   }
 
-  def markImportant(messageId: Int) = IsAuthenticated{ username => implicit request =>
+  def markImportant(messageId: Long) = IsAuthenticated{ username => implicit request =>
     //logger.info(s"in MessageController.markImportant(${messageId})")
     println(s"in MessageController.markImportant(${messageId})")
     
@@ -95,7 +95,7 @@ object MessageController extends Controller with Secured {
     Ok(HttpResponseUtil.success("Successfully marked important!"))
   }
 
-  def removeImportant(messageId: Int) = IsAuthenticated{ username => implicit request =>
+  def removeImportant(messageId: Long) = IsAuthenticated{ username => implicit request =>
     //logger.info(s"in MessageController.removeImportant(${messageId})")
     println(s"in MessageController.removeImportant(${messageId})")
     
@@ -103,7 +103,7 @@ object MessageController extends Controller with Secured {
     Ok(HttpResponseUtil.success("Successfully removed important!"))
   }
   
-  def markRead(messageId: Int) = IsAuthenticated{ username => implicit request =>
+  def markRead(messageId: Long) = IsAuthenticated{ username => implicit request =>
     //logger.info(s"in MessageController.markRead(${messageId})")
     println(s"in MessageController.markRead(${messageId})")
     
@@ -111,7 +111,7 @@ object MessageController extends Controller with Secured {
     Ok(HttpResponseUtil.success("Successfully marked read!"))
   }
   
-  def reply(messageId: Int) = IsAuthenticated(parse.json){ username => implicit request =>
+  def reply(messageId: Long) = IsAuthenticated(parse.json){ username => implicit request =>
     //logger.info(s"in MessageController.reply(${messageId})")
     println(s"in MessageController.reply(${messageId})")
     
@@ -133,7 +133,7 @@ object MessageController extends Controller with Secured {
     }
   }
   
-  def trash(messageId: Int) = IsAuthenticated{ username => implicit request =>
+  def trash(messageId: Long) = IsAuthenticated{ username => implicit request =>
     //logger.info(s"in MessageController.trash(${messageId})")
     println(s"in MessageController.trash(${messageId})")
     val optTrash = MessageBoxRepository.findTrash(userId)
@@ -146,7 +146,7 @@ object MessageController extends Controller with Secured {
     }
   }
   
-  def move(messageId: Int, messageBoxId: Int) = IsAuthenticated{ username => implicit request =>
+  def move(messageId: Long, messageBoxId: Long) = IsAuthenticated{ username => implicit request =>
     //logger.info(s"in MessageController.move(${messageId}, ${messageBoxId})")
     println(s"in MessageController.move(${messageId}, ${messageBoxId})")
     UserMessageRepository.moveMessageBox(messageId, messageBoxId, userId)
